Create auto-invoice-generate-email.py
This code uses invoice data stored in Excel to create invoices. There is an invoice template that is being applied to all invoices, made in Excel. We make an invoice for each unique company. Then we send the email to the email address associated with the company. I think this operates in O(n^2).